Extract WorldLocationStubbingMethods module
This is similar to what we've done with `FixtureMethods` and will make it easier
to re-use the methods in the regression tests.
